requirements.txt
langchain==0.3.25
langchain-openai==0.3.18
langgraph==0.4.1
openai==1.82.0
fastapi==0.115.12
uvicorn==0.34.2
supabase==2.15.1
pydantic==2.11.4
python-dotenv==1.1.0
pandas==2.2.3
httpx==0.28.1